bugfix: adjusted types of all entities

+++ src/main/java/com/panacademy/squad7/bluebank/domain/repositories/AddressesRepository.java
@@ -1,9 +1,10 @@
 package com.panacademy.squad7.bluebank.domain.repositories;
 
 import com.panacademy.squad7.bluebank.domain.models.Address;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AddressesRepository extends JpaRepository<Address, Integer> {
+public interface AddressesRepository extends JpaRepository<Address, Long> {
 }
