+++ src/main/java/com/panacademy/squad7/bluebank/exceptions/AddressNotFoundException.java
@@ -1,7 +1,7 @@
 package com.panacademy.squad7.bluebank.exceptions;
 
-public class AddressNotFoundException extends RuntimeException{
-    public AddressNotFoundException(Integer id){
+public class AddressNotFoundException extends RuntimeException {
+    public AddressNotFoundException(Long id) {
         super("Address not found with id " + id);
     }
 }
